+++ src/routes/strategies/[strategy=apiStrategy]/[status=positionStatus]-positions/[position=integer]/TradeTable.svelte
@@ -71,7 +71,7 @@
 </script>
 
 <section class="trade-table">
-	<h2>Rebalances</h2>
+	<h2>Trades</h2>
 	<DataTable {tableViewModel} targetableRows size="sm" />
 </section>